move documentStateFor() to services.js
